Add Mongoid::Client and add #with_session method to it.
